Refactor: merge all pig files into 1 file (pig.js)
